+++ src/hud/enchanter_screen.cpp
@@ -209,10 +209,10 @@ void _EnchanterScreen::RefreshBuyButtons() {
 		}
 		else if(ChildElement->Name == "button_max_skill_levels_buy") {
 			int CurrentLevel = HUD->Player->Character->MaxSkillLevels[SkillID];
-			int Cost = _Item::GetEnchantCost(CurrentLevel);
+			int64_t Cost = _Item::GetEnchantCost(CurrentLevel);
 
 			// Set button state
-			if(CurrentLevel >= Skill->MaxLevel || CurrentLevel >= Enchanter->Level || Cost > HUD->Player->Character->Attributes["Gold"].Int) {
+			if(CurrentLevel >= Skill->MaxLevel || CurrentLevel >= Enchanter->Level || Cost > HUD->Player->Character->Attributes["Gold"].Int64) {
 				ChildElement->SetEnabled(false);
 				ChildElement->Children.front()->Color = ae::Assets.Colors["gray"];
 			}
